% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RyggFigAndeler.R
\name{FigAndeler}
\alias{FigAndeler}
\title{Søylediagram som viser andeler av ulike variabler}
\usage{
FigAndeler(RegData, valgtVar, datoFra = "2007-01-01",
  datoTil = "2999-12-31", minald = 0, maxald = 130, erMann = "",
  hovedkat = 99, tittel = 1, outfile = "", reshID, enhetsUtvalg = 1)
}
\arguments{
\item{RegData}{En dataramme med alle nødvendige variabler fra registeret}

\item{valgtVar}{Hvilken variabel som skal visualiseres. Se \strong{Details} for oversikt.}

\item{datoFra}{Tidligste dato i utvalget (vises alltid i figuren).}

\item{datoTil}{Seneste dato i utvalget (vises alltid i figuren).}

\item{minald}{Alder, fra og med (Standardverdi: 0)}

\item{maxald}{Alder, til og med (Standardverdi: 130)}

\item{erMann}{Kjønn, standard: alt annet enn 0/1 gir begge kjønn
0: Kvinner
1: Menn}

\item{outfile}{Navn på fil figuren skrives til. Standard: '' (Figur skrives
til systemets standard utdataenhet (som regel skjerm))}

\item{reshID}{Parameter følger fra innlogging helseregister.no og angir
hvilken enhet i spesialisthelsetjenesten brukeren tilhører}

\item{enhetsUtvalg}{Gjør gruppeutvalg med eller uten sammenlikning for
 0: Hele landet
 1: Egen enhet mot resten av landet (Standard)
 2: Egen enhet
6: Egen enhet mot egen region 
7: Egen region 
8: Egen region mot resten}

\item{preprosess}{Preprosesser data
FALSE: Nei
TRUE: Ja (Standard)}

\item{hentData}{Gjør spørring mot database
0: Nei, RegData gis som input til funksjonen (Standard)
1: Ja}
}
\value{
Søylediagram (fordeling) av valgt variabel. De enkelte verdiene kan også sendes med.
}
\description{
Denne funksjonen lager et horisontalt eller vertikalt søylediagram som viser andeler (fordeling) 
av valgt variabel filtrert på de utvalg som er gjort.
}
\details{
Argumentet \emph{valgtVar} har følgende valgmuligheter:
   \itemize{
    \item Alder: Aldersfordeling 
    \item AntDagerInnl: Liggetid 
    \item Antibiotika: Er det gitt antibiotikaprofylakse?
    \item AntNivOpr: Antall nivå operert'
    \item ArbstatusPre: Arbeidsstatus før operasjon'
    \item Arbstatus3mnd: Arbeidsstatus 3 mnd. etter operasjon
    \item Arbstatus12mnd: Arbeidsstatus 12 mnd. etter operasjon
    \item ASA: ASA-grad
    \item BMI: Pasientenes BMI (Body Mass Index)
    \item EqangstPre: Helsetilstand: Angst
    \item EqgangePre: Helsetilstand: Gange
    \item ErstatningPre: Har pasienten søkt erstatning?
    \item Fornoyd3mnd: Fornøydhet 3 mnd etter operasjon
    \item Fornoyd12mnd: Fornøydhet 12 mnd etter operasjon
    \item HovedInngrep: Hovedinngrep
    \item Komorbiditet: Komorbiditet
    \item KomplPer: Peroperative komplikasjoner
    \item KomplPost: Pasientrapporterte komplikasjoner
    \item Liggedogn: Liggetid ved operasjon
    \item Morsmal: Morsmål
    \item Nytte3mnd: Hvilken nytte har du hatt av operasjonen? (svar 3 måneder etter)
    \item Nytte12mnd: Hvilken nytte har du hatt av operasjonen? (svar 12 måneder etter)
    \item OpInd: Operasjonsindikasjon
    \item OpIndPareseGrad: Operasjonsindikasjon, paresegrad
    \item OpIndSmeType: Operasjonsindikasjon, smertetype
    \item OpKat: Operasjonskategori 
    \item RadUnders: Radiologisk undersøkelse
    \item Roker: Røyker du?
    \item Saardren: Sårdren
    \item SivilStatus: Sivilstatus
    \item SmHyppPre: Hyppighet av smertestillende før operasjonen
    \item SmStiPre: Bruk av smertestillende før operasjonen
    \item SymptVarighRyggHof: Varighet av rygg-/hoftesmerter
    \item SympVarighUtstr: Varighet av utstrålende smerter
    \item TidlOpr: Tidligere ryggoperert?
    \item TidlOprAntall: Antall tidligere operasjoner
    \item UforetrygdPre: Har pasienten søkt uføretrygd?
    \item Underkat: Fordeling av inngrepstyper. NB: hovedkategori MÅ velges
    \item Utd: Høyeste fullførte utdanning
   }
}

