% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RyggFigGjsnTid.R
\name{GjsnTid}
\alias{GjsnTid}
\title{Utvikling over tid for gjennomsnitt/median av valgt variabel}
\usage{
GjsnTid(RegData, outfile, valgtVar, tidlOp = 0, erMann = "",
  hovedkat = 99, minald = 0, maxald = 130, ktr = 1, tittel = 1,
  datoFra = "2007-01-01", datoTil = "3000-01-01", valgtMaal = "",
  enhetsUtvalg = 1, hentData = 0, preprosess = 1, reshID)
}
\arguments{
\item{RegData}{En dataramme med alle nødvendige variabler fra registeret}

\item{outfile}{Navn på fil figuren skrives til. Standard: '' (Figur skrives
til systemets standard utdataenhet (som regel skjerm))}

\item{valgtVar}{Hvilken variabel som skal visualiseres. Se \strong{Details} for oversikt.}

\item{tidlOp}{Tidligere operert, 
       0: alle, 
       1: Tidl. operert samme nivå, 
       2: Tidl. operert annet nivå, 
3: Tidl. operert annet og sm. nivå,
4: Primæroperasjon}

\item{erMann}{Kjønn, standard: alt annet enn 0/1 gir begge kjønn
0: Kvinner
1: Menn}

\item{hovedkat}{Hvilken type hovedinngrep, numerisk 0-7, standard: 99, dvs. alle}

\item{minald}{Alder, fra og med (Standardverdi: 0)}

\item{maxald}{Alder, til og med (Standardverdi: 130)}

\item{ktr}{Hvilken oppfølging man ønsker å se på. Valg: 3mnd - 3 mnd kontroll(standard), 12mnd - 12 mnd kontroll}

\item{tittel}{Vise tittel i figuren eller ikke (0/1). standard:1}

\item{datoFra}{Tidligste operasjonsdato i utvalget (vises alltid i figuren).}

\item{datoTil}{Seneste operasjonsdato i utvalget (vises alltid i figuren).}

\item{valgtMaal}{'Gjsn': gir middelverdi (standard)
'Med': gir median}

\item{enhetsUtvalg}{Gjør gruppeutvalg med eller uten sammenlikning for
 0: Hele landet
 1: Egen enhet mot resten av landet (Standard)
 2: Egen enhet
6: Egen enhet mot egen region 
7: Egen region 
8: Egen region mot resten}

\item{hentData}{Gjør spørring mot database
0: Nei, RegData gis som input til funksjonen (Standard)
1: Ja}

\item{preprosess}{Preprosesser data
FALSE: Nei
TRUE: Ja (Standard)}

\item{reshID}{Parameter følger fra innlogging helseregister.no og angir
hvilken enhet i spesialisthelsetjenesten brukeren tilhører}
}
\value{
Linjediagram som viser utvikling over tid for valgt variabel
}
\description{
Figuren viser gjennomsnitt/median per år med konfidensintervall for valgt variabel.
I bakgrunn vises konfidensintervall for resten av landet.
}
\details{
Skala for de ulike livskvalitetsmålene er definert slik:
   \itemize{
	\item EQ5D: Skala fra -0.594 tl 1, jo høyere jo bedre.
	\item Oswestry: Skala fra 0 til 100, hvor lavest er friskest
    	\item Smerte: Skalaen går fra 0 til 10, dvs. at differansen ligger mellom -10 og 10..
   }
	
Argumentet \emph{valgtVar} har følgende valgmuligheter:
   \itemize{
	\item EQ5DEndr: Endring av EQ5D fra før til 3 eller 12 mnd. etter operasjonen 
	\item EQ5DPre: EQ5D før operasjon
    	\item Liggedogn: Liggetid på sykehuset
    	\item OswEndr: Endring i Oswestry-skår fra før til etter operasjon
    	\item OswTotPre: Oswestry-skår før operasjon
    	\item SmBeinEndr: Endring i beinsmerter fra før til etter operasjon
    	\item SmBePre: Grad av beinsmerter før operasjon
    	\item SmRyggEndr: Endring av ryggsmerter fra før til etter operasjon
    	\item SmRyPre: Ryggsmerter før operasjon
   }
}

