% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PverdiAndelsDiff.R
\name{PverdiAndelsDiff}
\alias{PverdiAndelsDiff}
\title{Beregn og juster p-verdi ved sammenligning av to og to andeler}
\usage{
PverdiAndelsDiff(n, N, konfnivaa = 0.95, justMetode = "fdr")
}
\arguments{
\item{n}{- Antall tellende hendelser, skalar eller vektor}

\item{N}{- Antall observasjoner, skalar eller vektor av samme lengde som n}

\item{konfnivaa}{- Konfidensnivået til test (standard 0.95)}
}
\value{
Justerte p-verdier for test av endring i andeler
}
\description{
Denne funksjonen benytter prop.test for å beregne p-verdi for at to andeler er like
og justerer p-verdiene ved p.adjust
}
