% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RyggFigAndeler.R
\name{RyggFigAndeler}
\alias{RyggFigAndeler}
\title{Søylediagram som viser andeler av ulike variabler}
\usage{
RyggFigAndeler(RegData, valgtVar, datoFra = "2007-01-01",
  datoTil = "2999-12-31", aar = 0, hentData = 0, preprosess = 1,
  minald = 0, maxald = 130, erMann = "", hovedkat = 99,
  opKat = 99, tidlOp = "", ktr = 0, tittelMed = 1, outfile = "",
  reshID = 0, enhetsUtvalg = 0)
}
\arguments{
\item{RegData}{En dataramme med alle nødvendige variabler fra registeret}

\item{valgtVar}{Hvilken variabel som skal visualiseres. Se \strong{Details} for oversikt.}

\item{datoFra}{Tidligste operasjonsdato i utvalget (vises alltid i figuren).}

\item{datoTil}{Seneste operasjonsdato i utvalget (vises alltid i figuren).}

\item{aar}{- Operasjonsår}

\item{hentData}{Gjør spørring mot database
0: Nei, RegData gis som input til funksjonen (Standard)
1: Ja}

\item{preprosess}{Preprosesser data
0: Nei
1: Ja (Standard)}

\item{minald}{Alder, fra og med (Standardverdi: 0)}

\item{maxald}{Alder, til og med (Standardverdi: 130)}

\item{erMann}{Kjønn, standard: alt annet enn 0/1 gir begge kjønn
0: Kvinner
1: Menn}

\item{hovedkat}{Hvilken type hovedinngrep, numerisk 0-9, standard: 99, dvs. alle. Valgmuligheter:
\itemize{
 \item 0:'Operasjonskategori: "ukjent"',    #hkat=0
 \item 1:'Prolapskirurgi',
 \item 2:'Foramenotomi',
 \item 3:'Laminektomi',
 \item 4:'Interspinøst implantat',    
 \item 5:'Fusjonskirurgi',
 \item 6:'Skiveprotese',
 \item 7:'Fjerning/rev. av implantat',
 \item 8:'Spinal stenose',
 \item 9:'Degen. spondylolistese'
 \item 99: Alle
 }}

\item{opKat}{Hastegrad av operasjon 1: Elektivt, 2: Akutt, 3: Halvøyeblikkelig. Slår sammen 1 og 3 til elektiv}

\item{tidlOp}{Tidligere operert, numerisk 1-4. Alt annet gir alle data uten utvalg.
       1: Tidl. operert samme nivå, 
       2: Tidl. operert annet nivå, 
3: Tidl. operert annet og sm. nivå,
4: Primæroperasjon}

\item{tittelMed}{Vise tittel i figuren eller ikke (0/1). standard:1}

\item{outfile}{Navn på fil figuren skrives til. Standard: '' (Figur skrives
til systemets standard utdataenhet (som regel skjerm))}

\item{reshID}{Parameter følger fra innlogging helseregister.no og angir
hvilken enhet i spesialisthelsetjenesten brukeren tilhører}

\item{enhetsUtvalg}{Gjør gruppeutvalg med eller uten sammenlikning. Se \strong{Details} for oversikt.}
}
\value{
Søylediagram (fordeling) av valgt variabel. De enkelte verdiene kan også sendes med.
}
\description{
Denne funksjonen lager et horisontalt eller vertikalt søylediagram som viser andeler (fordeling) 
av valgt variabel filtrert på de utvalg som er gjort.
}
\details{
Argumentet \emph{valgtVar} har følgende valgmuligheter:
   \itemize{
    \item alder: Aldersfordeling 
    \item antDagerInnl: Liggetid 
    \item antibiotika: Er det gitt antibiotikaprofylakse?
    \item antNivOpr: Antall nivå operert'
    \item arbstatus: Arbeidsstatus før, 3 el. 12 mnd. etter operasjon
    \item ASA: ASA-grad
    \item BMI: Pasientenes BMI (Body Mass Index)
    \item EQangstPre: Helsetilstand: Angst
    \item EQgangePre: Helsetilstand: Gange
    \item erstatningPre: Har pasienten søkt erstatning?
    \item fornoydhet: Fornøydhet 3 eller 12 mnd etter operasjon
    \item hovedInngrep: Hovedinngrep
    \item komorbiditet: Komorbiditet
    \item komplPer: Peroperative komplikasjoner
    \item komplPost: Pasientrapporterte komplikasjoner
    \item liggedogn: Liggetid ved operasjon
    \item morsmal: Morsmål
    \item nytte: Hvilken nytte har du hatt av operasjonen? (svar 3 eller 12 måneder etter)
    \item opInd: Operasjonsindikasjon
    \item opIndPareseGrad: Operasjonsindikasjon, paresegrad
    \item opIndSmeType: Operasjonsindikasjon, smertetype
    \item opKat: Operasjonskategori 
    \item radUnders: Radiologisk undersøkelse
    \item roker: Røyker du?
    \item saardren: Sårdren
    \item sivilStatus: Sivilstatus
    \item smHyppPre: Hyppighet av smertestillende før operasjonen
    \item smStiPre: Bruk av smertestillende før operasjonen
    \item SymptVarighRyggHof: Varighet av rygg-/hoftesmerter
    \item SymptVarighUtstr: Varighet av utstrålende smerter
    \item tidlOpr: Tidligere ryggoperert?
    \item tidlOprAntall: Antall tidligere operasjoner
    \item uforetrygdPre: Har pasienten søkt uføretrygd?
    \item underkat: Fordeling av inngrepstyper. NB: hovedkategori MÅ velges
    \item utd: Høyeste fullførte utdanning
   }
Argumentet \emph{hovedkat} har følgende valgmuligheter:
   \itemize{
    \item 0: Annet
    \item 1: Prolaps
    \item 2: Foramenotomi
    \item 3: Laminektomi
    \item 4: Eksp. intraspin. impl.
    \item 5: Fusjon
    \item 6: Skiveprotese
    \item 7: Fjerning/revisjon
   }
   Velges ingen av disse, vil alle data vises.
Argumentet \emph{enhetsUtvalg} har følgende valgmuligheter:
   \itemize{
    \item 0: Hele landet
    \item 1: Egen enhet mot resten av landet (Standard)
    \item 2: Egen enhet
    \item 3: Egen enhet mot egen sykehustype
    \item 4: Egen sykehustype
    \item 5: Egen sykehustype mot resten av landet
    \item 6: Egen enhet mot egen region
    \item 7: Egen region
   \item 8: Egen region mot resten
   	}
}
