% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RyggFigAndelTid.R
\name{RyggFigAndelTid}
\alias{RyggFigAndelTid}
\title{Tidstrend av andel opphold}
\usage{
RyggFigAndelTid(RegData, valgtVar, datoFra = "2011-01-01",
  datoTil = "3000-12-31", aar = 0, tidsenhet = "Aar", hovedkat = 99,
  ktr = 2, tidlOp = 99, tittel = 1, minald = 0, maxald = 130,
  erMann = 99, reshID = 0, outfile = "", opKat = 99, enhetsUtvalg = 1,
  preprosess = 1, hentData = 0, lagFig = 1, offData = 0)
}
\arguments{
\item{RegData}{En dataramme med alle nødvendige variabler fra registeret}

\item{valgtVar}{Hvilken variabel som skal visualiseres. Se \strong{Details} for oversikt.}

\item{datoFra}{Tidligste operasjonsdato i utvalget (vises alltid i figuren).}

\item{datoTil}{Seneste operasjonsdato i utvalget (vises alltid i figuren).}

\item{tidsenhet}{Oppløsning på tidsaksen. Verdier: 'Aar' (standard), 'Halvaar', 'Kvartal','Mnd'}

\item{hovedkat}{Hvilken type hovedinngrep, numerisk 0-7, standard: 99, dvs. alle}

\item{tidlOp}{Tidligere operert, numerisk 1-4. Alt annet gir alle data uten utvalg.
       1: Tidl. operert samme nivå, 
       2: Tidl. operert annet nivå, 
3: Tidl. operert annet og sm. nivå,
4: Primæroperasjon}

\item{tittel}{Vise tittel i figuren eller ikke (0/1). standard:1}

\item{minald}{Alder, fra og med (Standardverdi: 0)}

\item{maxald}{Alder, til og med (Standardverdi: 130)}

\item{erMann}{Kjønn, standard: alt annet enn 0/1 gir begge kjønn
0: Kvinner
1: Menn}

\item{reshID}{Parameter følger fra innlogging helseregister.no og angir
hvilken enhet i spesialisthelsetjenesten brukeren tilhører}

\item{outfile}{Navn på fil figuren skrives til. Standard: '' (Figur skrives
til systemets standard utdataenhet (som regel skjerm))}

\item{opKat}{Hastegrad av operasjon 1: Elektivt, 2: Akutt, 3: Halvøyeblikkelig}

\item{enhetsUtvalg}{Gjør gruppeutvalg med eller uten sammenlikning. Se \strong{Details} for oversikt.}

\item{preprosess}{Preprosesser data
0: Nei
1: Ja (Standard)}

\item{hentData}{Gjør spørring mot database
0: Nei, RegData gis som input til funksjonen (Standard)
1: Ja}
}
\value{
Figur som viser tidstrend, dvs. andel av valgt variabel for hvert år.
}
\description{
Denne funksjonen lager et søylediagram som viser andeler (fordeling) av valgt variabel
filtrert på de utvalg som er gjort.
}
\details{
Argumentet \emph{valgtVar} har følgende valgmuligheter:
   \itemize{
    \item alder70: Alder over 70år, 
    \item degSponFusj: Andel av degenerativ spondylolistese er operert  med fusjonskirurgi (hovedinngr=5)
    \item KpInf3Mnd: Sårinfeksjon, pasientrapportert
    \item OswEndr20: Forbedring av Oswestry-skår > 20 poeng
   }
}
