% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ER_ERSTATTET_RyggFigEndrPre.R
\name{RyggFigEndrPre}
\alias{RyggFigEndrPre}
\title{Punktplott med konfidensintervall. Viser gjennomsnittlig endring av valgt variabel vs. 
variabelens verdi før operasjonen.}
\usage{
RyggFigEndrPre(RegData, valgtVar, datoFra = "2007-01-01",
  datoTil = "2999-12-31", hentData = 0, preprosess = 1, minald = 0,
  maxald = 130, erMann = "", hovedkat = 99, tidlOp = "", tittel = 1,
  outfile = "", reshID, enhetsUtvalg = 1)
}
\arguments{
\item{RegData}{En dataramme med alle nødvendige variabler fra registeret}

\item{valgtVar}{Hvilken variabel som skal visualiseres. Se \strong{Details} for oversikt.}

\item{datoFra}{Tidligste operasjonsdato i utvalget (vises alltid i figuren).}

\item{datoTil}{Seneste operasjonsdato i utvalget (vises alltid i figuren).}

\item{hentData}{Gjør spørring mot database
0: Nei, RegData gis som input til funksjonen (Standard)
1: Ja}

\item{preprosess}{Preprosesser data
0: Nei
1: Ja (Standard)}

\item{minald}{Alder, fra og med (Standardverdi: 0)}

\item{maxald}{Alder, til og med (Standardverdi: 130)}

\item{erMann}{Kjønn, standard: alt annet enn 0/1 gir begge kjønn
0: Kvinner
1: Menn}

\item{hovedkat}{Hvilken type hovedinngrep, numerisk 0-7, standard: 99, dvs. alle}

\item{tidlOp}{Tidligere operert, numerisk 1-4. Alt annet gir alle data uten utvalg.
       1: Tidl. operert samme nivå, 
       2: Tidl. operert annet nivå, 
3: Tidl. operert annet og sm. nivå,
4: Primæroperasjon}

\item{tittel}{Vise tittel i figuren eller ikke (0/1). standard:1}

\item{outfile}{Navn på fil figuren skrives til. Standard: '' (Figur skrives
til systemets standard utdataenhet (som regel skjerm))}

\item{reshID}{Parameter følger fra innlogging helseregister.no og angir
hvilken enhet i spesialisthelsetjenesten brukeren tilhører}

\item{enhetsUtvalg}{Gjør gruppeutvalg med eller uten sammenlikning. Se \strong{Details} for oversikt.}
}
\value{
Søylediagram (fordeling) av valgt variabel. De enkelte verdiene kan også sendes med.
}
\description{
Gjennomsnitt per intervall med konfidensintervall for valgt variabel.
I bakgrunn vises konfidensintervall for resten av landet.
Også beregne uteliggere og ta disse med i utdata? I så fall bør de kunne spores tilbake til PID
}
\details{
Argumentet \emph{valgtVar} har følgende valgmuligheter:
   \itemize{
    \item EQ5DEndrPre: Skala fra -0.594 tl 1, jo høyere jo bedre.
    \item OswEndrPre: Oswestry (ODI-Oswestry Disability Index) Skala fra 0 til 100, hvor lavest er friskest.
    \item SmBeinEndrPre: Smerter i beina. Skalaen går fra 0 til 10, dvs. at differansen ligger mellom -10 og 10.
    \item SmRyggEndrPre: Smerter i ryggen Skalaen går fra 0 til 10, dvs. at differansen ligger mellom -10 og 10.
   }
Argumentet \emph{hovedkat} har følgende valgmuligheter:
   \itemize{
    \item 0: Annet
    \item 1: Prolaps
    \item 2: Foramenotomi
    \item 3: Laminektomi
    \item 4: Eksp. intraspin. impl.
    \item 5: Fusjon
    \item 6: Skiveprotese
    \item 7: Fjerning/revisjon
   }
   Velges ingen av disse, vil alle data vises.
Argumentet \emph{enhetsUtvalg} har følgende valgmuligheter:
   \itemize{
    \item 0: Hele landet
    \item 1: Egen enhet mot resten av landet (Standard)
    \item 2: Egen enhet
    \item 3: Egen enhet mot egen sykehustype
    \item 4: Egen sykehustype
    \item 5: Egen sykehustype mot resten av landet
    \item 6: Egen enhet mot egen region
    \item 7: Egen region
   \item 8: Egen region mot resten
   	}
}

