% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RyggFigGjsnBox.R
\name{RyggFigGjsnBox}
\alias{RyggFigGjsnBox}
\title{Utvikling over tid for gjennomsnitt/median av valgt variabel}
\usage{
RyggFigGjsnBox(RegData, outfile, valgtVar, tidlOp = "", erMann = "",
  hovedkat = 99, aar = 0, tidsenhet = "Mnd", opKat = 99,
  minald = 0, maxald = 130, ktr = 0, tittel = 1,
  datoFra = "2007-01-01", datoTil = "3000-01-01", valgtMaal = "Gjsn",
  enhetsUtvalg = 0, hentData = 0, preprosess = 1, reshID = 0)
}
\arguments{
\item{RegData}{En dataramme med alle nødvendige variabler fra registeret}

\item{outfile}{Navn på fil figuren skrives til. Standard: '' (Figur skrives
til systemets standard utdataenhet (som regel skjerm))}

\item{valgtVar}{Hvilken variabel som skal visualiseres}

\item{tidlOp}{Tidligere operert, numerisk 1-4. Alt annet gir alle data uten utvalg.
       1: Tidl. operert samme nivå, 
       2: Tidl. operert annet nivå, 
3: Tidl. operert annet og sm. nivå,
4: Primæroperasjon}

\item{erMann}{Kjønn, standard: alt annet enn 0/1 gir begge kjønn
0: Kvinner
1: Menn}

\item{hovedkat}{Hvilken type hovedinngrep, numerisk 0-9, standard: 99, dvs. alle. Valgmuligheter:
\itemize{
 \item 0:'Operasjonskategori: "ukjent"',    #hkat=0
 \item 1:'Prolapskirurgi',
 \item 2:'Foramenotomi',
 \item 3:'Laminektomi',
 \item 4:'Interspinøst implantat',    
 \item 5:'Fusjonskirurgi',
 \item 6:'Skiveprotese',
 \item 7:'Fjerning/rev. av implantat',
 \item 8:'Spinal stenose',
 \item 9:'Degen. spondylolistese'
 \item 99: Alle
 }}

\item{aar}{- Operasjonsår}

\item{opKat}{Hastegrad av operasjon 1: Elektivt, 2: Akutt, 3: Halvøyeblikkelig. Slår sammen 1 og 3 til elektiv}

\item{minald}{Alder, fra og med (Standardverdi: 0)}

\item{maxald}{Alder, til og med (Standardverdi: 130)}

\item{datoFra}{Tidligste operasjonsdato i utvalget (vises alltid i figuren).}

\item{datoTil}{Seneste operasjonsdato i utvalget (vises alltid i figuren).}

\item{valgtMaal}{'Gjsn': gir middelverdi (standard)
'Med': gir median}

\item{enhetsUtvalg}{Gjør gruppeutvalg med eller uten sammenlikning. Se \strong{Details} for oversikt.}

\item{hentData}{Gjør spørring mot database
0: Nei, RegData gis som input til funksjonen (Standard)
1: Ja}

\item{preprosess}{Preprosesser data
0: Nei
1: Ja (Standard)}

\item{reshID}{Parameter følger fra innlogging helseregister.no og angir
hvilken enhet i spesialisthelsetjenesten brukeren tilhører}
}
\value{
Linjediagram som viser utvikling over tid for valgt variabel
}
\description{
Figuren viser gjennomsnitt/median per år med konfidensintervall for valgt variabel.
I bakgrunn vises konfidensintervall for det man har valt å sammenlikne med.
}
\details{
Figurtypen dekker både utvikling over tid (gjsnTid) og sammenlikning med prescore (gjsnPre).

Argumentet \emph{valgtVar} har følgende valgmuligheter:
   \itemize{
	\item EQ5DEndr: Endring av EQ5D fra før til 3 eller 12 mnd. etter operasjonen 
	\item EQ5DPre: EQ5D før operasjon
         \item EQ5DEndrPre: EQ5D, endring vs. prescore
    	\item Liggedogn: Liggetid på sykehuset
    	\item OswEndr: Endring i Oswestry-skår fra før til etter operasjon
         \item OswEndrPre: Oswestry (ODI-Oswestry Disability Index.) Endring vs. prescore.
    	\item OswTotPre: Oswestry-skår før operasjon
    	\item SmBeinEndr: Endring i beinsmerter fra før til etter operasjon
         \item SmBeinEndrPre: Smerter i beina. Endring vs. prescore
    	\item SmBePre: Grad av beinsmerter før operasjon
    	\item SmRyggEndr: Endring av ryggsmerter fra før til etter operasjon
         \item SmRyggEndrPre: Smerter i ryggen. Endring vs. prescore 
    	\item SmRyPre: Ryggsmerter før operasjon
     \item TidOpReg MANGLER UTFYLT-variabel. (Ønsker å se på tid fra operasjon til registrering - )
   }
Argumentet \emph{enhetsUtvalg} har følgende valgmuligheter:
   \itemize{
    \item 0: Hele landet
    \item 1: Egen enhet mot resten av landet (Standard)
    \item 2: Egen enhet
    \item 3: Egen enhet mot egen sykehustype
    \item 4: Egen sykehustype
    \item 5: Egen sykehustype mot resten av landet
    \item 6: Egen enhet mot egen region 
    \item 7: Egen region 
   \item 8: Egen region mot resten [
   	}
}
