% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RyggPreprosesser.R
\name{RyggPreprosess}
\alias{RyggPreprosess}
\title{Preprosesser data fra Degenerativ Rygg}
\usage{
RyggPreprosess(RegData = RegData)
}
\arguments{
\item{RegData}{En dataramme med alle nødvendige variabler fra registeret}
}
\value{
RegData En dataramme med det preprosesserte datasettet
}
\description{
Denne funksjonen definerer og formaterer variabler
}
