% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RyggRegDataSQLFull.R
\name{RyggRegDataSQLFull}
\alias{RyggRegDataSQLFull}
\title{Henter data registrert for Degenerativ Rygg}
\usage{
RyggRegDataSQLFull()
}
\value{
RegData data frame
}
\description{
Henter data for Degenerativ Rygg fra "staging" (?)
}
