% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RyggUtvalg.R
\name{RyggUtvalg}
\alias{RyggUtvalg}
\title{Funksjon som gjør utvalg i datagrunnlaget til registreringene for Nakke}
\usage{
RyggUtvalg(RegData, datoFra, datoTil, minald = 0, maxald = 130,
  erMann = "", hovedkat = 99, aar = 0, opKat = 99, tidlOp = "",
  fargepalett = "BlaaOff")
}
\arguments{
\item{RegData}{En dataramme med alle nødvendige variabler fra registeret}

\item{datoFra}{Tidligste operasjonsdato i utvalget (vises alltid i figuren).}

\item{datoTil}{Seneste operasjonsdato i utvalget (vises alltid i figuren).}

\item{minald}{Alder, fra og med (Standardverdi: 0)}

\item{maxald}{Alder, til og med (Standardverdi: 130)}

\item{erMann}{Kjønn, standard: alt annet enn 0/1 gir begge kjønn
0: Kvinner
1: Menn}

\item{hovedkat}{Hvilken type hovedinngrep, numerisk 0-7, standard: 99, dvs. alle}

\item{opKat}{Hastegrad av operasjon 1: Elektivt, 2: Akutt, 3: Halvøyeblikkelig}

\item{tidlOp}{Tidligere operert, numerisk 1-4. Alt annet gir alle data uten utvalg.
       1: Tidl. operert samme nivå, 
       2: Tidl. operert annet nivå, 
3: Tidl. operert annet og sm. nivå,
4: Primæroperasjon}

\item{fargepalett}{- Velge fargepalett, standard:BlaaOff ("offentliggjøringsfargene")}
}
\description{
Funksjon som gjør utvalg av dataene, returnerer det reduserte datasettet og utvalgsteksten.
}
