% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AKjustering.R
\name{StandAlderKjonn}
\alias{StandAlderKjonn}
\title{Søylediagram med andeler for hver grupperingsenhet (sykehus, RHF, ...)}
\usage{
StandAlderKjonn(RegData, stdPop, antAldgr, katVariable)
}
\arguments{
\item{RegData}{E dataramme som minimum må inneholde variabelen det skal beregnes alders- og 
               kjønnsstandardiserte rater for. (variabelen: Variabel)
RegData inneholder både alder, kjønn og bo-områder.
Standardiserer mot variablene  Alder (år) og ErMann (0/1)
RegData må inneholde variabelen Variabel som det skal beregnes AK-justerte andeler for.}

\item{valgtVar}{Variabelen det skal vises resultat for. Se \strong{Details} for oversikt.}

\item{siste3aar}{0:Viser resultatat for hele perioden samlet, 1: Viser resultat for hvert av de siste tre år}

\item{AKjust}{Alders-og kjønnsjustering når grVar er boområder. Basert på 3 aldersgrupper gruppert ut fra alderskvartilene.
0:ikke juster, 1:juster for alder og kjønn}
}
\value{
Funksjon som alders- og kjønnsstandardiserer data basert på en valgt standardpopulasjon.
}
\description{
Funksjon som genererer en figur med andeler av en variabel for en valgt grupperingsvariabel,
f.eks. sykehus.
Funksjonen er delvis skrevet for å kunne brukes til andre grupperingsvariable enn sykehus
}
\details{
Andel som mottar sykepenger er definert som svaralternativene: 'Sykemeldt',
       'Aktiv sykemeldt', 'Delvis sykemeldt', 'Attføring/rehab.', 'Uføretrygdet'
}
